fix: move Loader component to Loader folder
